# model/openapi.yaml
openapi: 3.0.0
info:
  title: Microservice API
  description: API for managing microservices
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /microservices:
    get:
      summary: Get all microservices
      responses:
        '200':
          description: A list of microservices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Microservice'
    post:
      summary: Create a new microservice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Microservice'
      responses:
        '201':
          description: Microservice created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Microservice'

  /microservices/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The ID of the microservice to retrieve/update/delete
        schema:
          type: integer
          format: int64
    get:
      summary: Get a microservice by ID
      responses:
        '200':
          description: The microservice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Microservice'
        '404':
          description: Microservice not found
    put:
      summary: Update a microservice by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Microservice'
      responses:
        '200':
          description: Microservice updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Microservice'
        '404':
          description: Microservice not found
    delete:
      summary: Delete a microservice by ID
      responses:
        '204':
          description: Microservice deleted successfully
        '404':
          description: Microservice not found

  /microservices/search:
    get:
      summary: Search for microservices by name
      parameters:
        - name: serviceName
          in: query
          description: Microservice name to search for
          schema:
            type: string
      responses:
        '200':
          description: A list of microservices matching the search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Microservice'

components:
  schemas:
    Microservice:
      type: object
      properties:
        id:
          type: integer
          format: int64
        serviceName:
          type: string
        version:
          type: string
        description:
          type: string
        owner:
          type: string
        status:
          type: string
        dependencies:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        latestCommit:
          type: string
        gitRepoLink:
          type: string
      required:
        - serviceName
        - version
        - description
        - owner
        - status
        - dependencies
        - createdAt
        - updatedAt
        - latestCommit
        - gitRepoLink